reorganize pipeline dir and location of files.
add readmes to all dir.
