+++ README.md
@@ -6,8 +6,14 @@ Instead of remembering where to put everything and making a different choice for
 
 Think of it like a mini-pytorch lightening, with all the fory internals exposed for extension and modification.
 
+This project lives here: [https://github.com/publicmatt.com/ml_pipeline](https://github.com/publicmatt.com/ml_pipeline).
+
 
 # Usage
+
+```bash
+make help # lists available options.
+```
 
 ## Install:
 
@@ -17,10 +23,12 @@ Install the conda requirements:
 make install
 ```
 
-Which is a proxy for calling:
+## Data:
+
+Download mnist data from PJReadie's website:
 
 ```bash
-conda env updates -n ml_pipeline --file environment.yml
+make data
 ```
 
 ## Run:
@@ -31,7 +39,6 @@ Run the code on MNIST with the following command:
 make run
 ```
 
-
 # Tutorial
 
 The motivation for building a template for deep learning pipelines is this: deep learning is hard enough without every code baase being a little different.
@@ -41,33 +48,90 @@ Especially in a research lab, standardizing on a few components makes switching 
 In this template, you'll see the following:
 
 ## directory structure
+```
+.
+├── README.md
+├── environment.yml
+├── launch.sh
+├── Makefile
+├── data
+│   ├── mnist_test.csv
+│   └── mnist_train.csv
+├── docs
+│   └── 2023-01-26.md
+├── src
+│   ├── config
+│   │   └── main.yaml
+│   ├── data
+│   │   ├── __init__.py
+│   │   ├── README.md
+│   │   ├── collate.py
+│   │   └── dataset.py
+│   ├── eval.py
+│   ├── __init__.py
+│   ├── model
+│   │   ├── __init__.py
+│   │   ├── README.md
+│   │   ├── cnn.py
+│   │   └── linear.py
+│   ├── pipeline
+│   │   ├── __init__.py
+│   │   ├── README.md
+│   │   ├── logger.py
+│   │   ├── runner.py
+│   │   └── utils.py
+│   ├── sample.py
+│   └── train.py
+└── test
+    ├── __init__.py
+    └── test_pipeline.py
 
-- `src/model`
-- `src/config`
-- `data/`
-- `test/`
-    - pytest: unit testing.
-        - good for data shape
-        - TODO:
-- `docs/`
-    - switching projects is easier with these in place
-    - organize them
+8 directories, 25 files
 
-- `**/__init__.py`
-    - creates modules out of dir.
-    - `import module` works with these.
-- `README.md`
-    - root level required.
-    - can exist inside any dir.
+```
+
+## what and why?
+
 - `environment.yml`
-- `Makefile` 
+    - hutch research has standardized on conda
+    - here's a good tutorial on getting that setup: [seth email](emailto:bassetis@wwu.edu)
+- `launch.sh` or `Makefile`
     - to install and run stuff.
     - houses common operations and scripts.
-- `launch.sh` 
-    - script to dispatch training.
+    - `launch.sh` to dispatch training.
+- `README.md`
+    - explain the project and how to run it.
+    - list authors.
+    - list resources that new collaborators might need.
+    - root level dir.
+    - can exist inside any dir.
+    - reads nicely on github.com.
+- `docs/`
+    - switching projects is easier with these in place.
+    - organize them by meeting, or weekly agenda.
+    - generally collection of markdown files.
+- `test/`
+    - TODO
+    - pytest: unit testing.
+    - good for data shape. not sure what else.
+- `data/`
+    - raw data
+    - do not commit these to repo generally.
+        - `echo "*.csv" >> data/.gitignore`
+- `__init__.py`
+    - creates modules out of dir.
+    - `import module` works b/c of these.
+- `src/model/`
+    - if you have a large project, you might have multiple architectures/models.
+    - small projects might just have `model/VGG.py` or `model/3d_unet.py`.
+- `src/config`
+    - based on hydra python package.
+    - quickly change run variables and hyperparameters.
+- `src/pipeline`
+    - where the magic happens.
+    - `train.py` creates all the objects, hands them off to runner for batching, monitors each epoch.
 
 ## testing
-    
 - `if __name__ == "__main__"`.
     - good way to test things
 - enables lots breakpoints.
@@ -81,6 +145,8 @@ In this template, you'll see the following:
 
 ## data
 - collate functions!
+- datasets.
+- dataloader.
 
 ## formatting python
 - python type hints.
@@ -88,6 +154,8 @@ In this template, you'll see the following:
 
 ## running
 - tqdm to track progress.
+- wandb for logging.
 
 ## architecture
 - dataloader, optimizer, criterion, device, state are constructed in main, but passed to an object that runs batches.
+
